authority 表操作

+++ clerk-oauth/src/main/java/com/wind/clerk/oauth/service/impl/AuthorityServiceImpl.java
@@ -42,7 +42,10 @@ public class AuthorityServiceImpl implements AuthorityService {
     }
 
     @Override
-    public Boolean delete(Integer id) {
+    public Boolean delete(Integer id) throws Exception {
+        if (authorityMapper.countOfRelatedRoles(id) > 0) {
+            throw new Exception("无法删除,尚有关联角色");
+        }
         return authorityMapper.delete(id);
     }
 }
